+++ src/tests/unit_tests/sensor/utest_SEN_threadsafety.cpp
@@ -44,11 +44,12 @@ TEST(ChFilterAccess, data_access_safety) {
 
     auto cam = chrono_types::make_shared<ChCameraSensor>(
         box,                                                                // body camera is attached to
-        50,                                                                 // update rate in Hz
+        50.0f,                                                              // update rate in Hz
         chrono::ChFrame<double>({-8, 0, 1}, Q_from_AngAxis(0, {0, 1, 0})),  // offset pose
         1,                                                                  // image width
         1,                                                                  // image height
-        CH_C_PI / 3);                                                       // FOV
+        (float)CH_C_PI / 3                                                  // FOV
+    );                                                     
     cam->SetName("Camera Sensor");
     cam->PushFilter(chrono_types::make_shared<ChFilterRGBA8Access>());
     manager->AddSensor(cam);
